LayerSelector updated, switched to QListView/Model system
